requirements.txt
playwright==1.58.0
pandas>=2.0.0
openpyxl>=3.1.0
pyee>=11.1.0
greenlet>=3.0.0
gspread>=6.0.0
google-auth>=2.0.0
